DTO, Logica, Controller

+++ Fuentes/ServiceVentaComida/src/main/java/co/edu/usbcali/logica/PadreLogica.java
@@ -112,7 +112,7 @@ public class PadreLogica implements IPadreLogica {
 
 		@Override
 		@Transactional(readOnly=true)
-		public Padre consultarPorId(int padreId) {
+		public Padre consultarPorId(Long padreId) {
 			return padresDAO.consultarPorId(padreId);
 		}
